Threading working, basic client framework working.
Todo includes actually getting a server link, and using that to
define what our queue will be. That is for a future time.

Right now we have 1 master thread that starts a gui thread
which in turn runs two inter-dependant sub-threads.
The death of either sub-thread will trigger the death of the other.
Upon those sub-threads dying, the controller also dies.

The master is left to run the networking side of things.